+++ helper/terraformtype/helper/schema/type_resource.go
@@ -70,6 +70,37 @@ func (info *ResourceInfo) IsResource() bool {
 	return info.DeclaresField(ResourceFieldCreate)
 }
 
+// GetResourceMapResourceNames returns all resource names held in a map[string]*schema.Resource
+func GetResourceMapResourceNames(cl *ast.CompositeLit) []ast.Expr {
+	var result []ast.Expr
+
+	for _, elt := range cl.Elts {
+		switch v := elt.(type) {
+		case *ast.KeyValueExpr:
+			result = append(result, v.Key)
+		}
+	}
+
+	return result
+}
+
+// IsMapStringResource returns if the type is map[string]*Resource from the helper/schema package
+func IsMapStringResource(cl *ast.CompositeLit, info *types.Info) bool {
+	switch v := cl.Type.(type) {
+	case *ast.MapType:
+		switch k := v.Key.(type) {
+		case *ast.Ident:
+			if k.Name != "string" {
+				return false
+			}
+		}
+
+		return IsTypeResource(info.TypeOf(v.Value))
+	}
+
+	return false
+}
+
 // IsTypeResource returns if the type is Resource from the helper/schema package
 func IsTypeResource(t types.Type) bool {
 	switch t := t.(type) {